Move routes.js into api folder to fix Travis CI
